requirements.txt
pandas
numpy
pydantic
openai
instructor
tqdm
python-dotenv
nest-asyncio
datasets
huggingface_hub
matplotlib
pytest
